Integration clerk auth

+++ src/app/api/transaction/[id]/route.ts
@@ -1,3 +1,4 @@
+import { auth } from "@clerk/nextjs";
 import { NextRequest, NextResponse } from "next/server";
 import { prisma } from "~/utils/prisma";
 
@@ -7,6 +8,10 @@ interface Params {
   };
 }
 export async function DELETE(req: NextRequest, { params: { id } }: Params) {
-  await prisma.transaction.delete({ where: { id } });
+  const { userId } = auth();
+
+  if (!userId) return new Response("Unauthorized", { status: 401 });
+
+  await prisma.transaction.delete({ where: { id, clerkUserId: userId } });
   return NextResponse.json({ ok: true });
 }
